Moved IBlasExtensionsProvider functionality to IBlasProvider. There really is not any reason to divide this
