--- src/MGroup.LinearAlgebra/Implementations/IBlasExtensionsProvider.cs
@@ -1,15 +1,0 @@
-namespace MGroup.LinearAlgebra.Implementations
-{
-    /// <summary>
-    /// Provides linear algebra operations that are variations of the usual BLAS subroutines (see <see cref="IBlasProvider"/>) 
-    /// or in the same spirit.
-    /// Authors: Serafeim Bakalakos
-    /// </summary>
-    internal interface IBlasExtensionsProvider
-    {
-        /// <summary>
-        /// y = alpha * x + beta * y
-        /// </summary>
-        void Daxpby(int n, double alpha, double[] x, int offsetX, int incX, double beta, double[] y, int offsetY, int incY);
-    }
-}
