+++ src/MGroup.LinearAlgebra/Matrices/CscMatrix.cs
@@ -536,7 +536,7 @@ namespace MGroup.LinearAlgebra.Matrices
 					else
 					{
 						Array.Copy(this.values, resultValues, values.Length);
-						BlasExtensions.Daxpby(values.Length, otherCoefficient, otherCSC.values, 0, 1,
+						Blas.Daxpby(values.Length, otherCoefficient, otherCSC.values, 0, 1,
 							thisCoefficient, resultValues, 0, 1);
 					}
 					return new CscMatrix(NumRows, NumColumns, resultValues, this.rowIndices, this.colOffsets);
@@ -587,7 +587,7 @@ namespace MGroup.LinearAlgebra.Matrices
 			}
 			else
 			{
-				BlasExtensions.Daxpby(values.Length, otherCoefficient, otherMatrix.values, 0, 1, thisCoefficient, this.values, 0, 1);
+				Blas.Daxpby(values.Length, otherCoefficient, otherMatrix.values, 0, 1, thisCoefficient, this.values, 0, 1);
 			}
 		}
