+++ src/MGroup.LinearAlgebra/Matrices/Matrix.cs
@@ -959,7 +959,7 @@ namespace MGroup.LinearAlgebra.Matrices
 			else
 			{
 				Array.Copy(this.data, result, data.Length);
-				BlasExtensions.Daxpby(data.Length, otherCoefficient, otherMatrix.data, 0, 1, thisCoefficient, result, 0, 1);
+				Blas.Daxpby(data.Length, otherCoefficient, otherMatrix.data, 0, 1, thisCoefficient, result, 0, 1);
 			}
 			return new Matrix(result, NumRows, NumColumns);
 		}
@@ -1005,7 +1005,7 @@ namespace MGroup.LinearAlgebra.Matrices
 			}
 			else
 			{
-				BlasExtensions.Daxpby(data.Length, otherCoefficient, otherMatrix.data, 0, 1, thisCoefficient, this.data, 0, 1);
+				Blas.Daxpby(data.Length, otherCoefficient, otherMatrix.data, 0, 1, thisCoefficient, this.data, 0, 1);
 			}
 		}
