+++ src/MGroup.LinearAlgebra/Matrices/SymmetricMatrix.cs
@@ -522,7 +522,7 @@ namespace MGroup.LinearAlgebra.Matrices
 			else
 			{
 				Array.Copy(this.data, result, data.Length);
-				BlasExtensions.Daxpby(data.Length, otherCoefficient, otherMatrix.data, 0, 1, thisCoefficient, result, 0, 1);
+				Blas.Daxpby(data.Length, otherCoefficient, otherMatrix.data, 0, 1, thisCoefficient, result, 0, 1);
 			}
 			return new SymmetricMatrix(result, NumColumns, DefiniteProperty.Unknown);
 		}
@@ -555,7 +555,7 @@ namespace MGroup.LinearAlgebra.Matrices
 			}
 			else
 			{
-				BlasExtensions.Daxpby(data.Length, otherCoefficient, otherMatrix.data, 0, 1, thisCoefficient, this.data, 0, 1);
+				Blas.Daxpby(data.Length, otherCoefficient, otherMatrix.data, 0, 1, thisCoefficient, this.data, 0, 1);
 			}
 			this.Definiteness = DefiniteProperty.Unknown;
 		}
