+++ src/MGroup.LinearAlgebra/Matrices/TriangularLower.cs
@@ -370,7 +370,7 @@ namespace MGroup.LinearAlgebra.Matrices
             else
             {
                 Array.Copy(this.data, result, data.Length);
-                BlasExtensions.Daxpby(data.Length, otherCoefficient, otherMatrix.data, 0, 1, thisCoefficient, result, 0, 1);
+                Blas.Daxpby(data.Length, otherCoefficient, otherMatrix.data, 0, 1, thisCoefficient, result, 0, 1);
             }
             return new TriangularLower(result, NumRows);
         }
@@ -406,7 +406,7 @@ namespace MGroup.LinearAlgebra.Matrices
             }
             else
             {
-                BlasExtensions.Daxpby(data.Length, otherCoefficient, otherMatrix.data, 0, 1, thisCoefficient, this.data, 0, 1);
+                Blas.Daxpby(data.Length, otherCoefficient, otherMatrix.data, 0, 1, thisCoefficient, this.data, 0, 1);
             }
         }
