+++ tests/MGroup.LinearAlgebra.Implementations.NativeWin64.Tests/MGroup.LinearAlgebra.Implementations.NativeWin64.Tests.csproj
@@ -5,6 +5,7 @@
     <TargetFramework>net6.0</TargetFramework>
     <ImplicitUsings>enable</ImplicitUsings>
     <Nullable>enable</Nullable>
+    <StartupObject></StartupObject>
   </PropertyGroup>
 
   <ItemGroup>
