+++ tests/MGroup.LinearAlgebra.Implementations.NativeWin64.Tests/Program.cs
@@ -1,10 +1,7 @@
-﻿namespace MGroup.LinearAlgebra.Implementations.NativeWin64.Tests
+namespace MGroup.LinearAlgebra.Implementations.NativeWin64.Tests
 {
 	internal class Program
 	{
-		static void Main(string[] args)
-		{
-			Console.WriteLine("Hello, World!");
-		}
+		
 	}
 }
